upp.src: Fixed CLANG warnings in Win32
git-svn-id: svn://ultimatepp.org/upp/trunk@14099 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Core/mathutil.cpp
@@ -144,7 +144,7 @@ int64 ffloor64(double value)
 
 int64 fceil64(double value)
 {
-	return value >= INT64_MAX ? INT64_MAX : value <= INT64_MIN ? INT64_MIN
+	return value >= (double)INT64_MAX ? INT64_MAX : value <= (double)INT64_MIN ? INT64_MIN
 		: (int64)ceil(value);
 }
